==> tables/1_table-ones.Rmd <==
---
title: "EED-Development Draft Table 1"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  officedown::rdocx_document: 
    tables:
      style: Table Normal
  html_notebook: default
---

```{r setup, include = F}
rm(list = ls())

knitr::opts_chunk$set(echo = FALSE,
                      message = FALSE)

library(tidyverse)
library(kableExtra)
library(stringr)
library(tableone)
library(flextable)
library(officer)
library(here)
library(labelled)
library(gtsummary)
library(haven)

# CoopG/tableone2flextable
tableone2flextable <- function(tableone){
  rows <- nrow(tableone)
  cols <- ncol(tableone)
  rowsXcols <- rows*cols
  colnames <- colnames(tableone)
  
  listoflists <- list()
  
  for (i in 1:cols){
    start <- (i*rows+1)-rows
    end <- i*rows
    
    listoflists[[i]] <- tableone[start:end]
  }
  
  dataframe <- as.data.frame(listoflists, col.names = colnames)
  flex <- flextable::flextable(dataframe)
  
  return(flex)
}
```

```{r}
source(here("src/0_table-characteristics-prep.R"))
```

# Table 1. Characteristics of participants in Bangladesh

```{r, tab.cap = "Bangladesh"}

baselinechar_bg <- function(data){
  
  bg_child <- data %>% 
    mutate(across(c(ageday_st1, ageday_st2,
                    agedays_motor, agedays_cdi3),
                  ~ . / 30.5)) %>% 
    tbl_summary(include = c(sex, laz_t1, waz_t1, 
                            laz_t2, waz_t2,
                            diar7d_t2, diar7d_t3,
                            ageday_st1, ageday_st2,
                            agedays_motor, agedays_cdi3),
                type = list(sex ~ "dichotomous",
                            c(laz_t1, waz_t1,
                              laz_t2, waz_t2,
                              ageday_st1, ageday_st2,
                              agedays_motor, agedays_cdi3) ~ "continuous"),
                value = list(sex ~ "female"),
                digits = list(all_categorical() ~ c(0, 2),
                              all_continuous() ~ c(2, 2, 2)),
                missing = "no")
  
  bg_mom <- data %>% 
    distinct(dataid, hhid, 
             momage, momheight_raw,
             cesd_sum_t2, cesd_sum_ee_t3,
             pss_sum_mom_t3, momedu) %>% 
    tbl_summary(include = c(momage, momheight_raw,
                            cesd_sum_t2, cesd_sum_ee_t3,
                            pss_sum_mom_t3, momedu),
                digits = list(all_categorical() ~ c(0, 2),
                              all_continuous() ~ c(2, 2, 2),
                              cesd_sum_t2 ~ 0),
                type = list(c(momage,
                              cesd_sum_t2,
                              cesd_sum_ee_t3) ~ "continuous"),
                missing = "no")
  
  bg_house <- data %>% 
    distinct(dataid, hhid, 
             Ncomp, Nlt18,
             has_elec, 
             has_floor, has_roof, 
             has_asset_phone, has_asset_moto,
             hfiacat, 
             tr) %>% 
    tbl_summary(include = c(Ncomp, Nlt18,
                            has_elec, 
                            has_floor, has_roof, 
                            has_asset_phone, has_asset_moto,
                            hfiacat, 
                            tr),
                label = list(Ncomp ~ "Number of people per compound", 
                             Nlt18 ~ "Number of children <18 years in the household",
                             has_elec ~ "Has electricity", 
                             has_floor ~ "Has a cement floor",
                             has_roof ~ "Has an iron roof", 
                             has_asset_phone ~ "Owns a mobile phone",
                             has_asset_moto ~ "Owns a motorcycle",
                             hfiacat ~ "Household food insecurity"),
                type = list(Nlt18 ~ "continuous"),
                digits = list(all_categorical() ~ c(0, 2),
                              all_continuous() ~ c(2, 2, 2),
                              c(Ncomp, Nlt18) ~ 0),
                missing = "no")
  
  return(
    list(
      "bg_child" = bg_child,
      "bg_mom"   = bg_mom, 
      "bg_house" = bg_house
    )
  )
}


tbl_all_bg <- bg %>% 
  baselinechar_bg()

tbl_all_bg %>% 
  tbl_stack(group_header = c("Child", "Mother", "Household")) %>%
  modify_footnote_body(footnote = "Any level of food insecurity assessed using the Household Food Insecurity Access Scale",
                       columns = "label",
                       rows = variable == "hfiacat" & row_type == "label") %>% 
  modify_footnote_body(footnote = "CESD-R = Center for Epidemiologic Studies Depression Scale Revised",
                       columns = "label",
                       rows = str_detect(variable, "cesd_") & row_type == "label") %>% 
  as_gt() %>%
  gt::tab_style(
    style = gt::cell_text(weight = "bold"),
    locations = gt::cells_row_groups(groups = everything())
  ) 
```


\* 

** Any level of food insecurity assessed using the Household Food Insecurity Access Scale 


\newpage

# Table 2. Characteristics of participants in Kenya

```{r, tab.cap = "Kenya"}
baselinechar_kenya <- function(k_data){
  k_child <- k_data %>% 
    mutate(across(c(aged1, aged2,
                    agedays_motor, childage_dev),
                  ~ . / 30.5)) %>% 
    tbl_summary(include = c(sex, laz_t1, waz_t1, 
                            laz_t2, waz_t2,
                            # diarr7_t0, diarr7_t1,
                            aged1, aged2,
                            agedays_motor, childage_dev),
                type = list(sex ~ "dichotomous"#,
                            # diarr7_t0 ~ "dichotomous",
                            # diarr7_t1 ~ "dichotomous"
                ),
                value = list(sex ~ "female"#,
                             # diarr7_t0 ~ 1,
                             # diarr7_t1 ~ 1
                ),
                digits = list(all_categorical() ~ c(0, 2),
                              all_continuous() ~ c(2, 2, 2)),
                missing = "no")
  
  k_mom <- k_data %>% 
    distinct(hhid, momage, momheight_raw,
             phq_score_t3,
             pss_score, momedu) %>% 
    tbl_summary(include = c(momage, momheight_raw,
                            phq_score_t3,
                            pss_score, momedu),
                digits = list(all_categorical() ~ c(0, 2),
                              all_continuous() ~ c(2, 2, 2)),
                missing = "no")
  
  k_house <- k_data %>% 
    distinct(hhid, Ncomp, Nlt18,
             has_electricity, 
             has_floor, has_roof, 
             has_mobile, has_motorcycle,
             hhs_mod_sev,
             tr) %>% 
    tbl_summary(include = c(Ncomp, Nlt18,
                            has_electricity, 
                            has_floor, has_roof, 
                            has_mobile, has_motorcycle,
                            hhs_mod_sev,
                            tr),
                label = list(Ncomp ~ "Number of people per compound", 
                             Nlt18 ~ "Number of children <18 years in the household",
                             has_electricity ~ "Has electricity", 
                             has_floor ~ "Has a cement floor",
                             has_roof ~ "Has an iron roof", 
                             has_mobile ~ "Owns a mobile phone",
                             has_motorcycle ~ "Owns a motorcycle"),
                type = list(Nlt18 ~ "continuous"), 
                digits = list(all_categorical() ~ c(0, 2),
                              all_continuous() ~ c(2, 2, 2),
                              c(Ncomp, Nlt18) ~ 0 ),
                missing = "no")
  
  list("k_child" = k_child,
       "k_mom" = k_mom,
       "k_house" = k_house) 
}

tbl_all_k <- k %>% 
  baselinechar_kenya() 

tbl_all_k %>% 
  tbl_stack(group_header = c("Child", "Mother", "Household")) %>%
  modify_footnote_body(footnote = "Moderate to severe hunger defined using the Household Hunger Scale",
                       columns = "label",
                       rows = variable == "hhs_mod_sev" & row_type == "label") %>% 
  modify_footnote_body(footnote = "PHQ = Patient Health Questionnaire",
                       columns = "label",
                       rows = str_detect(variable, "phq_") & row_type == "label") %>% 
  as_gt() %>%
  gt::tab_style(
    style = gt::cell_text(weight = "bold"),
    locations = gt::cells_row_groups(groups = everything())
  ) 


```


\* PHQ = Patient Health Questionnaire 

** Moderate to severe hunger defined using the Household Hunger Scale
